+++ Project17-C-Map/InteractiveClusteringMap/Map/MapViewController.swift
@@ -121,44 +121,5 @@ final class MapViewController: UIViewController {
         }
     }
     
-extension MapViewController: NMFMapViewCameraDelegate {
-    func mapView(_ mapView: NMFMapView, cameraIsChangingByReason reason: Int) {
-        DispatchQueue.main.async { [weak self] in
-            guard let self = self else { return }
-            
-            self.transparentLayer?.sublayers?.forEach { subLayer in
-                self.setMarkerPosition(marker: subLayer)
-            }
-        }
-    }
-    
-    func mapView(_ mapView: NMFMapView, cameraWillChangeByReason reason: Int, animated: Bool) {
-        DispatchQueue.main.async { [weak self] in
-            guard let self = self else { return }
-            
-            self.transparentLayer?.sublayers?.forEach { subLayer in
-                self.setMarkerPosition(marker: subLayer)
-            }
-        }
-    }
-    
-    func mapView(_ mapView: NMFMapView, cameraDidChangeByReason reason: Int, animated: Bool) {
-        DispatchQueue.main.async { [weak self] in
-            guard let self = self else { return }
-            
-            self.transparentLayer?.sublayers?.forEach { subLayer in
-                self.setMarkerPosition(marker: subLayer)
-            }
-        }
-    }
-    
-    func mapViewCameraIdle(_ mapView: NMFMapView) {
-        DispatchQueue.main.async { [weak self] in
-            guard let self = self else { return }
-            
-            self.transparentLayer?.sublayers?.forEach { subLayer in
-                self.setMarkerPosition(marker: subLayer)
-            }
-        }
-    }
 }
+
